Fixing branch based on feedback
