+++ QA/generic_tests.py
@@ -30,53 +30,6 @@ def qa_test_table_updated(table, db, **kwargs):
         raise Exception(f"{db}.{table} WAS NOT UPDATED FOR THE CURRENT TIMEFRAME")
 
 
-def update_to_granular_version_indicator(table, db):
-    from lib.configuration import get_current_config, get_connection_string
-    config = get_current_config(type=db, **{"execution_date": datetime.date(2000, 1, 1)})
-    cstr = get_connection_string(config, 'PROD_DB')
-    engine = create_engine(cstr)
-    if db == 'granted_patent':
-        id = 'id'
-        fk = 'patent_id'
-        fact_table = 'patent'
-    else:
-        id = 'document_number'
-        fk = 'document_number'
-        fact_table = 'publication'
-    query = f"""
-update {table} update_table 
-	inner join {fact_table} p on update_table.{fk}=p.{id}
-set update_table.version_indicator=p.version_indicator     
-    """
-    print(query)
-    query_start_time = time()
-    engine.execute(query)
-    query_end_time = time()
-    print("This query took:", query_end_time - query_start_time, "seconds")
-
-def update_version_indicator(table, db, **kwargs):
-    from lib.configuration import get_current_config, get_connection_string
-    config = get_current_config(type=db, schedule="quarterly", **kwargs)
-    ed = process_date(config['DATES']["end_date"], as_string=True)
-    cstr = get_connection_string(config, 'PROD_DB')
-    engine = create_engine(cstr)
-    query = f"""
-update {table} update_table 
-set update_table.version_indicator={ed} 
-    """
-    print(query)
-    query_start_time = time()
-    engine.execute(query)
-    query_end_time = time()
-    print("This query took:", query_end_time - query_start_time, "seconds")
-
-
 
 if __name__ == '__main__':
-    update_to_granular_version_indicator('cpc_current', 'pgpubs')
-    # update_version_indicator('lawyer', 'granted_patent', **{
-    #     "execution_date": datetime.date(2021, 10, 1)
-    # })
-    # update_version_indicator('persistent_assignee_disambig_long', 'granted_patent', **{
-    #     "execution_date": datetime.date(2021, 10, 1)
-    # })
+    print()
